server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "blinker==1.9.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "dotenv==0.9.9",
    "flask==3.1.0",
    "flask-cors==5.0.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.166.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.4",
    "googleapis-common-protos==1.69.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "httplib2==0.22.0",
    "idna==3.10",
    "flask-login>=0.6.3",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "markupsafe==3.0.2",
    "numpy==2.1.2",
    "opencv-python>=4.11.0.86",
    "pandas==2.2.3",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.2",
    "pydantic-core==2.33.1",
    "pymongo==4.11.3",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2024.2",
    "requests==2.32.3",
    "rsa==4.9",
    "pandas==2.2.3",
    "pymongo==4.11.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2024.2",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "six==1.16.0",
    "threadpoolctl==3.5.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.1",
    "typing-inspection==0.4.0",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "tzdata==2024.2",
    "werkzeug==3.1.3",
    "pillow",
    "pyzbar>=0.1.9",
]
